refactor: @CurrentUser 파일 디렉토리 변경
