+++ server/momo-api/src/main/java/com/momo/api/auth/CurrentUserArgumentResolver.java
@@ -3,7 +3,6 @@ package com.momo.api.auth;
 import java.util.Objects;
 import javax.servlet.http.HttpServletRequest;
 
-import com.momo.common.CurrentUser;
 import com.momo.domain.auth.service.OAuthService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.core.MethodParameter;
