+++ server/momo-api/src/main/java/com/momo/api/group/GroupController.java
@@ -1,6 +1,6 @@
 package com.momo.api.group;
 
-import com.momo.common.CurrentUser;
+import com.momo.api.auth.CurrentUser;
 import com.momo.domain.common.dto.EnumResponse;
 import com.momo.domain.group.dto.GroupCreateRequest;
 import com.momo.domain.group.dto.GroupResponse;
