+++ server/momo-api/src/main/java/com/momo/api/post/CommentController.java
@@ -1,6 +1,6 @@
 package com.momo.api.post;
 
-import com.momo.common.CurrentUser;
+import com.momo.api.auth.CurrentUser;
 import com.momo.domain.post.dto.CommentCreateRequest;
 import com.momo.domain.post.dto.CommentResponse;
 import com.momo.domain.post.service.CommentService;
